Se cambia la clase 'Moneda' por la clase 'Segell' para verficar el correcto funcionamiento del contenedor generico, se realizan puebas - OK - Hay que mejorar algunos metodos

+++ src/main/java/clases/Contenedor.java
@@ -106,7 +106,7 @@ public class Contenedor<E extends Comparable<E> & InterfaceDataAlfabetic> implem
     }
 
     // Para iterar la lista generica por un valor ingresado
-    public Iterator<E> iterarPorValor(int minVal) {
+    public Iterator<E> iterarPorValor(double minVal) {
         return new Iterator<E>() {
             int indice = 0; // servira para poder recorrer el arrayList
